[Laboratory 3]

+++ university-course-manager/src/main/java/pl/edu/pg/eti/kask/ucm/university/service/impl/UniversityServiceImpl.java
@@ -3,6 +3,8 @@ package pl.edu.pg.eti.kask.ucm.university.service.impl;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
 import lombok.NoArgsConstructor;
+import pl.edu.pg.eti.kask.ucm.course.entity.Course;
+import pl.edu.pg.eti.kask.ucm.course.repository.api.CourseRepository;
 import pl.edu.pg.eti.kask.ucm.university.entity.University;
 import pl.edu.pg.eti.kask.ucm.university.repository.api.UniversityRepository;
 import pl.edu.pg.eti.kask.ucm.university.service.api.UniversityService;
@@ -17,9 +19,13 @@ public class UniversityServiceImpl implements UniversityService {
 
     private final UniversityRepository universityRepository;
 
+    private final CourseRepository courseRepository;
+
     @Inject
-    public UniversityServiceImpl(UniversityRepository universityRepository) {
+    public UniversityServiceImpl(UniversityRepository universityRepository,
+                                 CourseRepository courseRepository) {
         this.universityRepository = universityRepository;
+        this.courseRepository = courseRepository;
     }
 
     @Override
@@ -49,6 +55,18 @@ public class UniversityServiceImpl implements UniversityService {
 
     @Override
     public void delete(UUID id) {
-        this.universityRepository.delete(id);
+        Optional<University> universityOpt = this.find(id);
+
+        if (universityOpt.isPresent()) {
+            University university = universityOpt.get();
+
+            List<Course> courses = this.courseRepository.findAllByUniversity(university);
+
+            for (Course course : courses) {
+                this.courseRepository.delete(course.getId());
+            }
+
+            this.universityRepository.delete(id);
+        }
     }
 }
